refactor(types): rename PictureSource to ISource and update properties for consistency
