+++ src/runtime/components/BoosterVimeo/props.ts
@@ -1,4 +1,4 @@
-import type { PictureSource } from '../../../module';
+import type { ISource } from '../../../module';
 
 export default {
   url: {
@@ -30,7 +30,7 @@ export default {
   },
 
   posterSources: {
-    type: Array<PictureSource>,
+    type: Array<ISource>,
     default() {
       return [
         {
